removed temp files + moved bench over mock to internal/xsql/bench
